Update Adaptive UI code sequence (organization only)
